update
-fixes memory leaks
-quits if client.dylib or main pid not found
-waits if glowobject not found until found
-type "stop" or "exit" to terminate properly
-automatically terminates when csgo_osx64 terminates
