+++ Wall/MemMngr.hpp
