--- Wall/memory/manager.hpp
@@ -1,16 +1,0 @@
-//
-//  manager.hpp
-//
-//  Created by Andre Kalisch on 27.09.16.
-//  Copyright © 2016 Andre Kalisch. All rights reserved.
-//
-
-#ifndef memmngr_hpp
-#define memmngr_hpp
-
-struct Temp_t
-{
-    char buffer[256];
-};
-
-#endif /* memmngr_hpp */
