--- Wall/memory/process.hpp
@@ -1,23 +1,0 @@
-//
-//  process.hpp
-//
-//  Created by sheep on 16/5/14.
-//  Copyright © 2016年 nadp. All rights reserved.
-//
-
-#ifndef Utils_hpp
-#define Utils_hpp
-
-#include <stdio.h>
-#include <stdlib.h>
-#include <mach/mach_traps.h>
-#include <mach/mach_init.h>
-#include <mach/mach_error.h>
-#include <mach/mach.h>
-#include <mach-o/dyld_images.h>
-#include <mach-o/loader.h>
-#include <libproc.h>
-#include <sys/stat.h>
-#include <string>
-
-#endif /* Utils_hpp */
